refactor(api/plant): split calcGrowthStage() into plantService
TodoService.complete()にPlantの成長ロジックが混在していたため、
成長段階計算の責務をPlantServiceへ分離する。
型定義(GrowthStageResult)も合わせてplant.types.tsへ移動。
